+++ yafsrc/YAFNET.Core/Services/UpgradeService.cs
@@ -212,6 +212,13 @@ public class UpgradeService(IServiceLocator serviceLocator, IRaiseEvent raiseEve
 
             migrator.Run();
         }
+
+        if (prevVersion < 1001)
+        {
+            var migrator = new Migrator(this.DbAccess.ResolveDbFactory(), typeof(Migration1001));
+
+            migrator.Run();
+        }
     }
 
     /// <summary>