+++ yafsrc/YAFNET.Types/InputModels/EditBoardInputModel.cs
@@ -42,6 +42,12 @@ public class EditBoardInputModel
     public string Name { get; set; }
 
     /// <summary>
+    /// Gets or sets the description.
+    /// </summary>
+    /// <value>The description.</value>
+    public string Description { get; set; }
+
+    /// <summary>
     /// Gets or sets the culture.
     /// </summary>
     /// <value>The culture.</value>