+++ yafsrc/YAFNET.Types/InputModels/SettingsInputModel.cs
@@ -36,6 +36,12 @@ public class SettingsInputModel
     public string Name { get; set; }
 
     /// <summary>
+    /// Gets or sets the description.
+    /// </summary>
+    /// <value>The description.</value>
+    public string Description { get; set; }
+
+    /// <summary>
     /// Gets or sets the forum email.
     /// </summary>
     /// <value>The forum email.</value>